fix(testing): use int64_t for __ry_testing__reportFail line parameter
Ry's `int` lowers to `i64` in IR, so the @native("testing") dispatch
emits `call void @__ry_testing__reportFail(i64 %line, ptr %msg)`. The
C signature `int line` (i32) was an ABI mismatch — works on x86-64 /
ARM64 for line numbers fitting in 32 bits but violates strict C++ ABI
rules.

Take `int64_t` to match the IR call, then narrow to `int` only at the
`__ry_test_fail` boundary so its legacy 32-bit signature stays
out-of-scope for #718.

Addresses CodeRabbit review on PR #1632.

+++ src/runtime_testing.cpp
@@ -1,9 +1,16 @@
 #include "ry/test_runtime.hpp"
 
+#include <cstdint>
+
 namespace ry {
 
-extern "C" void __ry_testing__reportFail(int line, const char *msg) {
-    __ry_test_fail(line, msg);
+// Ry's `int` lowers to `i64` in IR, so the @native("testing") dispatch
+// emits `call void @__ry_testing__reportFail(i64 %line, ptr %msg)`.  The
+// C signature must therefore receive `int64_t` to match the ABI.  We
+// narrow to `int` for the existing `__ry_test_fail` helper, which keeps
+// its legacy 32-bit line parameter (line numbers fit comfortably).
+extern "C" void __ry_testing__reportFail(int64_t line, const char *msg) {
+    __ry_test_fail(static_cast<int>(line), msg);
 }
 
 } // namespace ry
